+++ CHANGELOG.md
@@ -2,6 +2,7 @@ See also http://pvp.haskell.org/faq
 
 ### 0.1.6.0
 
+* New function `openConnectionSSL'` which allows to customize the 'SSL' connection /before/ a client SSL handshake is attempted.
 * New convenience function `getContextSSL` function allowing to retrieve global `SSLContext`.
 
 ### 0.1.5.0